refactor: sms 디렉토리 global 디렉토리로 이동
